Moved and cleaned up the code a bit
