+++ testcontainers-base/src/main/java/no/vegvesen/ixn/docker/keygen/generator/RootCAKeyGenerator.java
@@ -1,4 +1,4 @@
-package no.vegvesen.ixn.docker;
+package no.vegvesen.ixn.docker.keygen.generator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -44,10 +44,6 @@ public class RootCAKeyGenerator  extends GenericContainer<RootCAKeyGenerator> {
 
     public Path getCaCertOnHost() {
         return keysFolder.resolve(String.format("ca.%s.crt.pem",caDomain));
-    }
-
-    public RootCaDetails getRootCaDetails() {
-        return new RootCaDetails(getCaKeyOnHost(),getCaCertOnHost());
     }
 
 }
