+++ testcontainers-base/src/test/java/no/vegvesen/ixn/docker/RootCAKeyGeneratorIT.java
@@ -1,5 +1,6 @@
 package no.vegvesen.ixn.docker;
 
+import no.vegvesen.ixn.docker.keygen.generator.RootCAKeyGenerator;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
