+++ testcontainers-base/src/test/java/no/vegvesen/ixn/docker/ServiceProviderCertGeneratorIT.java
@@ -1,5 +1,6 @@
 package no.vegvesen.ixn.docker;
 
+import no.vegvesen.ixn.docker.keygen.generator.ServiceProviderCertGenerator;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
